+++ bamParsing.py
@@ -2,35 +2,6 @@ from bamnostic import AlignmentFile
 import streamlit as st
 
 from fullRead import FullRead
-
-#A class to hold all attributes of a read for easier aggregate analysis
-# class FullRead:
-#     def __init__(self, name, sequence, qualities, mods = []):
-#         self.name = name
-#         self.sequence = sequence
-#         self.qualities = qualities
-#         self.mods = mods
-    
-#     #Uses the mod scores to create a u containing sequence
-#     def generate_U_sequence(self, seq, mods, U_thresh):
-#         new_seq = []
-#         i = 0
-#         for base in seq:
-#             if base != "T":
-#                 new_seq.append(base)
-#             else:
-#                 next_mod = mods[i]
-#                 i += 1
-#                 if next_mod >= U_thresh:
-#                     new_seq.append("U")
-#                 else:
-#                     new_seq.append("T")
-#         return new_seq
-
-#     #Uses the mod scores and a threshold to see how many uracils are suspected in a read
-#     def get_U_count(self, thresh):
-#         return sum(1 for mod in self.mods if mod >= thresh)
-    
                     
 #Extract the name, sequence, qscores, and mods from a read and put them in a FullRead object
 def get_data_from_read(bam_path, target_read_name) -> tuple[FullRead, int]: